Simplify & inference fns moved under mathics.eval (#1140)
